+++ graphix/command.py
@@ -4,24 +4,19 @@ from __future__ import annotations
 
 import dataclasses
 import enum
+import logging
 from enum import Enum
 from typing import ClassVar, Literal
 
 from graphix import utils
-from graphix.clifford import Clifford
-from graphix.fundamentals import (
-    ANGLE_PI,
-    Angle,
-    ParameterizedAngle,
-    Plane,
-    Sign,
-)
-from graphix.measurements import Domains
-from graphix.pauli import Pauli
+from graphix.clifford import Clifford, Domains
+from graphix.measurements import Measurement
 from graphix.repr_mixins import DataclassReprMixin
 from graphix.states import BasicStates, State
 
 Node = int
+
+logger = logging.getLogger(__name__)
 
 
 class CommandKind(Enum):
@@ -112,18 +107,15 @@ class M(BaseM, _KindChecker):
     ----------
     node : int
         Node index of the measured qubit.
-    plane : Plane, optional
-        Measurement plane, defaults to :class:`~graphix.fundamentals.Plane.XY`.
-    angle : ParameterizedAngle, optional
-        Rotation angle divided by :math:`\pi`.
+    measurement : Measurement
+        Measurement description.
     s_domain : set[int], optional
         Domain for the X byproduct operator.
     t_domain : set[int], optional
         Domain for the Z byproduct operator.
     """
 
-    plane: Plane = Plane.XY
-    angle: ParameterizedAngle = 0
+    measurement: Measurement = Measurement.X
     s_domain: set[Node] = dataclasses.field(default_factory=set)
     t_domain: set[Node] = dataclasses.field(default_factory=set)
     kind: ClassVar[Literal[CommandKind.M]] = dataclasses.field(default=CommandKind.M, init=False)
@@ -138,15 +130,13 @@ class M(BaseM, _KindChecker):
 
         Returns
         -------
-        :class:`~graphix.command.M`
+        :class:`M`
             Equivalent command representing the pattern ``MC``.
         """
         domains = clifford_gate.commute_domains(Domains(self.s_domain, self.t_domain))
-        update = MeasureUpdate.compute(self.plane, False, False, clifford_gate)
         return M(
             self.node,
-            update.new_plane,
-            self.angle * update.coeff + update.add_term,
+            self.measurement.clifford(clifford_gate),
             domains.s_domain,
             domains.t_domain,
         )
@@ -249,60 +239,3 @@ class T(_KindChecker, BaseCommand):
 
 Command = N | M | E | C | X | Z | S | T
 Correction = X | Z
-
-
-@dataclasses.dataclass
-class MeasureUpdate:
-    r"""Describe how a measure is changed by signals and a vertex operator.
-
-    Parameters
-    ----------
-    new_plane : Plane
-        Updated measurement plane after commuting gates.
-    coeff : int
-        Coefficient by which the angle is multiplied.
-    add_term : Angle
-        Additional term to add to the measurement angle.
-    """
-
-    new_plane: Plane
-    coeff: int
-    add_term: Angle
-
-    @staticmethod
-    def compute(plane: Plane, s: bool, t: bool, clifford_gate: Clifford) -> MeasureUpdate:
-        r"""Compute the measurement update.
-
-        Parameters
-        ----------
-        plane : ~graphix.fundamentals.Plane
-            Measurement plane of the command.
-        s : bool
-            Whether an :math:`X` signal is present.
-        t : bool
-            Whether a :math:`Z` signal is present.
-        clifford_gate : ~graphix.clifford.Clifford
-            Vertex operator applied before the measurement.
-
-        Returns
-        -------
-        MeasureUpdate
-            Update describing the new measurement.
-        """
-        gates = list(map(Pauli.from_axis, plane.axes))
-        if s:
-            clifford_gate = Clifford.X @ clifford_gate
-        if t:
-            clifford_gate = Clifford.Z @ clifford_gate
-        gates = list(map(clifford_gate.measure, gates))
-        new_plane = Plane.from_axes(*(gate.axis for gate in gates))
-        cos_pauli = clifford_gate.measure(Pauli.from_axis(plane.cos))
-        sin_pauli = clifford_gate.measure(Pauli.from_axis(plane.sin))
-        exchange = cos_pauli.axis != new_plane.cos
-        coeff = -1 if exchange == (cos_pauli.unit.sign == sin_pauli.unit.sign) else 1
-        add_term: Angle = 0
-        if cos_pauli.unit.sign == Sign.MINUS:
-            add_term += ANGLE_PI
-        if exchange:
-            add_term = ANGLE_PI / 2 - add_term
-        return MeasureUpdate(new_plane, coeff, add_term)